Extract common ConationsSection component
